Added HiveKillList.pm This is a hive version of the KillList module, it's more or less the same with the exception of it not reading from the killlist config file. Instead it gets the config information directly from the corresponding analysis hash in the pipeline config file. For example the following is from the analysis hash of an exonerate analysis: USE_KILL_LIST       => 1, KILL_TYPE           => 'protein', KILL_LIST_FILTER    => {                          -only_mol_type        => 'protein',                          -user_id              => undef,                          -from_source_species  => undef,                          -before_date          => undef,                          -having_status        => undef,                          -reasons              => [],                          -for_analyses         => [],                          -for_species          => [],                          -for_external_db_ids  => [],                        },
It may vary from module to module (hopefully not though), but in the above example the info is set in the config_settings
hash that is within the main analysis hash. The database details from the killlist are defined along with the rest of the
databases at the start of the pipeline config file. Other than the change in the way of passing in the information the module
should work identically to the non-hive version